+++ RELEASES.md
@@ -9,9 +9,13 @@ Some of the changes done in this version is:
 - Added support for mobile devices 
 - Added devicePixelRatio support to look sharp on high-resolution screens
 - Added info/statistics button in menu
+- Added export button in menu, clicking will export Timeline in opposite format from opened JSON --> CSV | CSV --> JSON
+- Added skipEmptyLines when parsing CSV-files
 - Fixed focus bug in modal window
 - Fixed date bug when parsing CSV-files
 - Fixed pan-functions to work as desired in all browsers
+- Fixed left-margin for button(s) in Alert dialog
+- Improved the overall code quality
 
 ## 1.1.0
 2022-08-03